+++ poetry/console/commands/debug/__init__.py
@@ -1,1 +1,2 @@
+from .info import DebugInfoCommand
 from .resolve import DebugResolveCommand